# src/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexport ERP API",
    "version": "1.0.0",
    "description": "Automatically generated from all Express routes in the Nexport ERP backend."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/uploads/{filename}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/uploads/{filename}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/branding/logo/{mode}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "mode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/branding/logos": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/dashboard/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/shipments/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shippingLine",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "originCountry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "originPort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destinationPort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "containerType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "containerSize",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "blNumber": {
                    "example": "any"
                  },
                  "bookingNumber": {
                    "example": "any"
                  },
                  "shippingLine": {
                    "example": "any"
                  },
                  "freightForwarder": {
                    "example": "any"
                  },
                  "vesselName": {
                    "example": "any"
                  },
                  "voyageNumber": {
                    "example": "any"
                  },
                  "etd": {
                    "example": "any"
                  },
                  "eta": {
                    "example": "any"
                  },
                  "actualArrival": {
                    "example": "any"
                  },
                  "originCountry": {
                    "example": "any"
                  },
                  "originPort": {
                    "example": "any"
                  },
                  "destinationPort": {
                    "example": "any"
                  },
                  "warehouseLocation": {
                    "example": "any"
                  },
                  "deliveryAddress": {
                    "example": "any"
                  },
                  "priority": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "currency": {
                    "example": "any"
                  },
                  "companyId": {
                    "example": "any"
                  },
                  "exporterCompanyId": {
                    "example": "any"
                  },
                  "tags": {
                    "example": "any"
                  },
                  "internalNotes": {
                    "example": "any"
                  },
                  "goodsDescription": {
                    "example": "any"
                  },
                  "notes": {
                    "example": "any"
                  },
                  "notificationUserIds": {
                    "example": "any"
                  },
                  "carrierTrackingLastEvent": {
                    "example": "any"
                  },
                  "isActive": {
                    "example": "any"
                  },
                  "containers": {
                    "example": "any"
                  },
                  "shipmentItems": {
                    "example": "any"
                  },
                  "items": {
                    "example": "any"
                  },
                  "requiredDocumentIds": {
                    "example": "any"
                  },
                  "invoiceNumber": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/containers/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "containerType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "containerSize",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shipmentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkedShipment",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "containerNumber": {
                    "example": "any"
                  },
                  "containerType": {
                    "example": "any"
                  },
                  "containerSize": {
                    "example": "any"
                  },
                  "sealNumber": {
                    "example": "any"
                  },
                  "stuffingType": {
                    "example": "any"
                  },
                  "weightCapacity": {
                    "example": "any"
                  },
                  "currentWeight": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "currentLocation": {
                    "example": "any"
                  },
                  "latitude": {
                    "example": "any"
                  },
                  "longitude": {
                    "example": "any"
                  },
                  "shipmentId": {
                    "example": "any"
                  },
                  "goodsDescription": {
                    "example": "any"
                  },
                  "isActive": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/companies/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "contactPerson": {
                    "example": "any"
                  },
                  "mobile": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "officeAddress": {
                    "example": "any"
                  },
                  "gstNumber": {
                    "example": "any"
                  },
                  "iecCode": {
                    "example": "any"
                  },
                  "panNumber": {
                    "example": "any"
                  },
                  "bankName": {
                    "example": "any"
                  },
                  "bankAccount": {
                    "example": "any"
                  },
                  "bankIfsc": {
                    "example": "any"
                  },
                  "billingAddress": {
                    "example": "any"
                  },
                  "shippingAddress": {
                    "example": "any"
                  },
                  "creditLimit": {
                    "example": "any"
                  },
                  "companyType": {
                    "example": "any"
                  },
                  "type": {
                    "example": "any"
                  },
                  "taxNumber": {
                    "example": "any"
                  },
                  "bankDetails": {
                    "example": "any"
                  },
                  "customFields": {
                    "example": "any"
                  },
                  "otherDetails": {
                    "example": "any"
                  },
                  "isActive": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/exporter-companies/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "contactPerson": {
                    "example": "any"
                  },
                  "mobile": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "address": {
                    "example": "any"
                  },
                  "country": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "countryOfOrigin",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "category": {
                    "example": "any"
                  },
                  "hsCode": {
                    "example": "any"
                  },
                  "sku": {
                    "example": "any"
                  },
                  "brandName": {
                    "example": "any"
                  },
                  "unitType": {
                    "example": "any"
                  },
                  "countryOfOrigin": {
                    "example": "any"
                  },
                  "companyId": {
                    "example": "any"
                  },
                  "isActive": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/brand-logos": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Administrator session token returned by the authentication API."
      },
      "CronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "The CRON_SECRET environment variable."
      }
    }
  }
}